pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.3.2",
    "pysocks>=1.7.1",
    "python-dotenv>=0.19.2",
    "requests>=2.28.0",
    "telethon>=1.24.0",
]
